Split widget creation page to hooks and components
