+++ src/app/[locale]/account/team/[teamId]/(withLeftNav)/integrations/widgets/[widgetId]/components/WidgetConfigForm/index.ts
@@ -1,0 +1,3 @@
+import WidgetConfigForm from "./WidgetConfigForm";
+
+export default WidgetConfigForm;
