+++ src/app/[locale]/account/team/[teamId]/(withLeftNav)/integrations/widgets/[widgetId]/const.ts
@@ -1,0 +1,6 @@
+export enum TabValues {
+    CONFIGURATION = "configuration",
+    PREVIEW = "preview",
+}
+
+export const DATA_CUSTODIAN_LIMIT = 25;
